Modify tests so that a moderator group can be tested, clean up the entries tests a bit
